Refactored deployer code to unexport stuff.
Previously, `impl.KubeConfig`, `impl.BuildAndUploadDockerImage`, and
`impl.GenerateYAMLs` were all exported and used by
`cmd/weaver-kube/deploy.go`. This PR moves the `deploy.go` code into
`internal/impl/deploy.go` to unexport these things, and instead exports
a single `Deploy` function.

I also moved all user-provided config into a `config.go` file.
